recovery: nand-squashfs: Bring in block-based OTA upgrade support
squashfs-on-nand depends on gluebi-emulated mtdblock-devices.
gluebi is an additional driver which emulates MTD devices on
top of UBI volumes; for each UBI volume, an MTD device is created,
and all I/O to this MTD device is redirected to the UBI volume.

Most of the heavy lifting to support block-based OTA upgrade
on nand-squashfs is done at the time of package generation.
A bootup script that runs in recovery-mode (before recovery
is launched) symlinks /dev/block/bootdevice/by-name/<partition>
to the mtdblock devices. The only change needed in recovery
is to increase the maximum number of MTD-partition entries mtd-utils
can hold (the previous limit of 32 partitions is surpassed due to
additional MTD devices being created by gluebi).

Change-Id: I8cee26c4cf701d7c529e131c3d2f96bcc9a47a09