Fix -Wreturn-type warnings as much as possible
With GCC 8.2 upgrade, I observed aborts in recovery due to
stack smashing.
After turning on -Werror, most of the errors were due to no return
statement in functions that return non-void. In C, 
this would be fine but it is an undefined-behavior in C++,
which in our case was resulting in aborts due to stack-smash.

As it turns out, GCC 8 enforces the assumption that control 
never reaches the end of a non-void function in C++.
https://gcc.gnu.org/gcc-8/porting_to.html



Test: 1) With these fixes no aborts observed in recovery.
      2) With -Werror=return-type, no compilation errors
         observed.
Change-Id: I94c286673042143da4693b2d824e1588fea415be